Temporary workaround to emulate old GDK-API of poppler.
This is a temporary workaround. GDK-API has been removed from poppler in
poppler 0.17 (commit 149b7fe) but pdfpres relies on that API. So for
now, I just copied the old poppler code. Of course, this workaround is
likely to break.

Unfortunately, I'm running out of free time. If you'd like to help me
improve pdfpres, please send me patches and/or fork the project on
GitHub. I'd very much appreciate it! :-)