version: Allow building without Git.
Whoops. I blindly assumed that everyone is building from inside a Git
repository. This may not be the case.

When building without Git, show a simple string. I will have to adjust
that string after the next "release".

Taken as a whole, the "version" feature got a lot more complicated than
I initially thought... Sorry for the premature push to GitHub.